# docs/openapi-mvp.yaml
openapi: 3.0.3
info:
  title: Open Movie Planer API
  version: 0.1.0
  description: MVP API fuer Projekt-, Budget-, Kosten- und Drehplanung.

servers:
  - url: http://localhost:3000

tags:
  - name: Projects
  - name: CostCenters
  - name: Expenses
  - name: ShootDays
  - name: Dashboard

paths:
  /projects:
    get:
      tags: [Projects]
      summary: Listet Projekte
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags: [Projects]
      summary: Erstellt ein Projekt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: Holt Projekt-Details
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Not Found
    put:
      tags: [Projects]
      summary: Aktualisiert ein Projekt
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      tags: [Projects]
      summary: Loescht ein Projekt
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '204':
          description: No Content

  /projects/{projectId}/cost-centers:
    get:
      tags: [CostCenters]
      summary: Listet Kostenstellen eines Projekts
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CostCenter'
    post:
      tags: [CostCenters]
      summary: Erstellt eine Kostenstelle
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCostCenterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CostCenter'

  /projects/{projectId}/expenses:
    get:
      tags: [Expenses]
      summary: Listet Ausgaben eines Projekts
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Expense'
    post:
      tags: [Expenses]
      summary: Erfasst eine Ausgabe
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Expense'

  /projects/{projectId}/shoot-days:
    get:
      tags: [ShootDays]
      summary: Listet Drehtage eines Projekts
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShootDay'
    post:
      tags: [ShootDays]
      summary: Legt einen Drehtag an
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShootDayRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShootDay'

  /projects/{projectId}/dashboard:
    get:
      tags: [Dashboard]
      summary: Liefert aggregierte Dashboard-Daten fuer ein Projekt
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDashboard'

components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pre, production, post]
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
      required: [id, title, status]

    CreateProjectRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pre, production, post]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required: [title, status]

    UpdateProjectRequest:
      allOf:
        - $ref: '#/components/schemas/CreateProjectRequest'

    CostCenter:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        budget:
          type: number
          format: double
      required: [id, projectId, name, budget]

    CreateCostCenterRequest:
      type: object
      properties:
        name:
          type: string
        budget:
          type: number
          format: double
      required: [name]

    Expense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        costCenterId:
          type: string
          format: uuid
          nullable: true
        amount:
          type: number
          format: double
        description:
          type: string
          nullable: true
        expenseDate:
          type: string
          format: date
          nullable: true
      required: [id, projectId, amount]

    CreateExpenseRequest:
      type: object
      properties:
        costCenterId:
          type: string
          format: uuid
        amount:
          type: number
          format: double
        description:
          type: string
        expenseDate:
          type: string
          format: date
      required: [amount]

    ShootDay:
      type: object
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        date:
          type: string
          format: date
        location:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [id, projectId, date]

    CreateShootDayRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        location:
          type: string
        notes:
          type: string
      required: [date]

    DashboardCostCenter:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        budget:
          type: number
          format: double
        spent:
          type: number
          format: double
        delta:
          type: number
          format: double
      required: [id, name, budget, spent, delta]

    ProjectDashboard:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
        totalBudget:
          type: number
          format: double
        totalSpent:
          type: number
          format: double
        remainingBudget:
          type: number
          format: double
        nextShootDays:
          type: array
          items:
            $ref: '#/components/schemas/ShootDay'
        topCostCenters:
          type: array
          items:
            $ref: '#/components/schemas/DashboardCostCenter'
      required: [projectId, totalBudget, totalSpent, remainingBudget, nextShootDays, topCostCenters]